carousel and bootstrap
I made some changes on the code that might affect other styling or script code. I changed: - class “dots” to “dot” in index.html
“circles” was changed to “dots”
in the JS file I changed most of the code. The carousel changes picture every 4 seconds, when you hover the mouse it stops and when moveout the mouse it continues.
I added a logo and a cart image(save in the file)
the bootstrap link is now at the bottom of the index page ( this way it does not interferes with the images in the carousel.)